pyproject.toml
[project]
name = "hyperstat"
version = "0.1.0"
description = "Hyperliquid stat-arb bot (backtest + paper + live) with funding overlay"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Hippolyte Del Rio" }]
license = { text = "Proprietary" }

dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "scipy>=1.11",
  "pyyaml>=6.0",
  "pydantic>=2.6",
  "python-dotenv>=1.0",
  "httpx>=0.27",
  "websockets>=12.0",
  "tenacity>=8.2",
  "rich>=13.7",
  "typer>=0.12",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "ruff>=0.6",
  "mypy>=1.8",
  "types-PyYAML>=6.0.12.20241230",
]

# Optionnel: si tu veux utiliser le SDK officiel Hyperliquid plus tard
hyperliquid = [
  "hyperliquid-python-sdk>=0.10.0"
]

[project.scripts]
hyperstat = "hyperstat.main:app"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
extend-select = ["I"]  # import sorting

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
